requirements.txt
fastapi==0.89.1
numpy==1.23.4
pandas==1.5.1
psycopg2==2.9.5
pydantic==1.10.4
requests==2.28.1
SQLAlchemy==1.4.46
toloka_kit==0.1.23
